+++ test/controller_test.rb
@@ -1,0 +1,8 @@
+require_relative "test_helper"
+
+class ControllerTest < ActionDispatch::IntegrationTest
+  def test_index
+    get blazer.root_path
+    assert_response :success
+  end
+end