+++ test/test_helper.rb
@@ -1,0 +1,13 @@
+require "bundler/setup"
+require "combustion"
+Bundler.require(:default)
+require "minitest/autorun"
+require "minitest/pride"
+
+logger = ActiveSupport::Logger.new(ENV["VERBOSE"] ? STDERR : nil)
+
+Combustion.path = "test/internal"
+Combustion.initialize! :active_record, :action_controller, :sprockets do
+  config.action_controller.logger = logger
+  config.active_record.logger = logger
+end